Add fixed up version of bt-agent
The bt-agent package in Debian doesn't correspond to any version of the
bluez-tools on its new GitHub home. So either the tarball that used to
be on Google Code was modified compared to the checked in version, or
the Debian packager uploaded a modified tarball.

In the meantime, and for the sake of expediency, ship a binary for the
CHIP, built from https://github.com/hadess/bluez-tools