Disable video output in Linux
This means that connecting to the device via USB will be necessary
to configure it further if required, but should significantly reduce the
noise during normal usage.

Note that I couldn't figure out how to disable the video output in
u-boot and in the kernel before the video arguments is parsed.

Closes: #5